+++ 1859856/Calculadora/calculadora.h
@@ -1,0 +1,21 @@
+#indef__CALCULADORA__H
+#define__CALCULADORA__H
+
+class calculadora {
+	
+	float suma
+	float resta
+	float multiplicacion
+	float division
+	
+	
+};
+
+    float suma();
+	float resta();
+	float multiplicacion;
+	float division;
+	
+	
+#indef
+
