--- 1859856/calculadora.h
@@ -1,21 +1,0 @@
-#indef__CALCULADORA__H
-#define__CALCULADORA__H
-
-class calculadora {
-	
-	float suma
-	float resta
-	float multiplicacion
-	float division
-	
-	
-};
-
-    float suma();
-	float resta();
-	float multiplicacion;
-	float division;
-	
-	
-#indef
-
